move lock token function into keeper
